Application selection page:
fixed - The store state for the application Id now retrieves the application that has been selected.

fixed - added history.js and also used it in the router switching the import from
BrowserRouter as Router to Router (in order to make the onclick() function call)
Link would navigate before even updating state.

Needs Work - the styling needs to be added. Code comments need to be cleaned up a little.